backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
requests
pandas
numpy
scikit-learn
tensorflow
python-multipart 
python-jose[cryptography]
passlib[bcrypt]
catboost
langchain
langchain-community
langchain-openai
langchain-huggingface
faiss-cpu
sentence-transformers
joblib
pillow
